requirements.txt
# Agri-AI Inference Service Dependencies
# ===========================================

# Core dependencies
inference-sdk>=0.56.0
flask>=2.0.0

# Additional useful packages for production
gunicorn>=20.0.0
python-dotenv>=0.19.0
requests>=2.25.0
pillow>=8.0.0

# Development dependencies (optional)
pytest>=6.0.0
pytest-flask>=1.2.0
black>=21.0.0
flake8>=3.8.0